Versão do professor para fazer a última parte

+++ src/main/java/br/edu/ulbra/election/voter/service/VoterService.java
@@ -1,15 +1,12 @@
 package br.edu.ulbra.election.voter.service;
 
-import br.edu.ulbra.election.voter.client.VoteClientService;
+import br.edu.ulbra.election.voter.client.ElectionClientService;
 import br.edu.ulbra.election.voter.exception.GenericOutputException;
 import br.edu.ulbra.election.voter.input.v1.VoterInput;
 import br.edu.ulbra.election.voter.model.Voter;
 import br.edu.ulbra.election.voter.output.v1.GenericOutput;
-import br.edu.ulbra.election.voter.output.v1.VoteOutput;
 import br.edu.ulbra.election.voter.output.v1.VoterOutput;
 import br.edu.ulbra.election.voter.repository.VoterRepository;
-import feign.FeignException;
-
 import org.apache.commons.lang.StringUtils;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
@@ -29,18 +26,17 @@ public class VoterService {
 
     private final PasswordEncoder passwordEncoder;
 
-    private final VoteClientService voteClientService;
-    
+    private final ElectionClientService electionClientService;
+
     private static final String MESSAGE_INVALID_ID = "Invalid id";
     private static final String MESSAGE_VOTER_NOT_FOUND = "Voter not found";
-    private static final String MESSAGE_VOTER_VOTED = "Voter already voted";
-    
+
     @Autowired
-    public VoterService(VoterRepository voterRepository, ModelMapper modelMapper, PasswordEncoder passwordEncoder, VoteClientService voteClientService){
+    public VoterService(VoterRepository voterRepository, ModelMapper modelMapper, PasswordEncoder passwordEncoder, ElectionClientService electionClientService){
         this.voterRepository = voterRepository;
         this.modelMapper = modelMapper;
         this.passwordEncoder = passwordEncoder;
-        this.voteClientService = voteClientService;
+        this.electionClientService = electionClientService;
     }
 
     public List<VoterOutput> getAll(){
@@ -100,22 +96,17 @@ public class VoterService {
         if (voter == null){
             throw new GenericOutputException(MESSAGE_VOTER_NOT_FOUND);
         }
-        
-        try {
-            List<VoteOutput> votes = voteClientService.getByVoterId(voterId);
-            if (votes.size() > 0) {
-            	throw new GenericOutputException(MESSAGE_VOTER_VOTED);
-            }
-        } catch (FeignException e) {
-        	if (e.status() == 500) {
-                throw new GenericOutputException(MESSAGE_VOTER_NOT_FOUND);
-            }
+
+        GenericOutput genericOutput = electionClientService.findVotesByVoter(voter.getId());
+        if (!genericOutput.getMessage().equals("0")){
+            throw new GenericOutputException("Voter already voted");
         }
-        
+
         voterRepository.delete(voter);
 
         return new GenericOutput("Voter deleted");
     }
+
     private void checkEmailDuplicate(String email, Long currentVoter){
         Voter voter = voterRepository.findFirstByEmail(email);
         if (voter != null && !voter.getId().equals(currentVoter)){
